% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViceCountyCheck.R
\name{CompareCollectionCensus}
\alias{CompareCollectionCensus}
\title{Closest Vice County - Bryophytes}
\usage{
CompareCollectionCensus(SpecimenColumn, VCColumn)
}
\arguments{
\item{SpecimenColumn}{The column where the specimen names are located in the dataset}

\item{VCColumn}{The column where the vice county is located in the dataset}
}
\value{
Dataframe ViceCountyDifferences with list of species, the vice county that is different and the closest vice county
}
\description{
This function uses the 2021 British Bryological Society Census Catalogue. For each species get a list of the vice county that is not in the census data and the closest vice county to that (numerically) from the census data.
}
\examples{
CompareCollectionCensus(ExampleData$ScientificName, ExampleData$WatsonianViceCounty)
}
