% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViceCountyCheck.R
\name{DistributionMap}
\alias{DistributionMap}
\title{Species Distribution Maps With Bryophyte Census Data}
\usage{
DistributionMap(Record = 1, Directory)
}
\arguments{
\item{Record}{The record number for the species in the data to start producing maps for. Default = 1.}

\item{Directory}{Directory of shapefile}
}
\value{
Plots of species distribution maps from collection data and from census catalogue data. Maps are made one from user selection in console.
}
\description{
Produce species distribution heat maps for vice counties in Britain from the species in SuspectSpecies alongside distribution maps from census catalogue data.
}
\examples{
DistributionMap(Directory = "")
}
