% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViceCountyCheck.R
\name{DistributionMapNoCensus}
\alias{DistributionMapNoCensus}
\title{Species Distribution Maps Without Census Data}
\usage{
DistributionMapNoCensus(Record = 1)
}
\arguments{
\item{Record}{The record number for the species in the data to start producing maps for. Default = 1.}
}
\value{
Plots of species distribution maps from collection data. Maps are made one from user selection in console.
}
\description{
Produce species distribution heat maps for vice counties in Britain from the species in SuspectSpecies.
}
\examples{
DistributionMapNoCensus()
}
