% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example specimen dataset}
\format{
A data frame with 48 rows and 5 variables
\describe {
 \item{ScientificName}{Specimen scientific name}
 \item{WatsonianViceCounty}{Specimen vice county}
 \item{Collector1}{First collector}
 \item{Collector2}{Second collector}
 \item{Collector3}{Third collector}
}
}
\source{
Created in house to serve as an example
}
\usage{
ExampleData
}
\description{
Includes example specimens with ScientificName and WatsonianViceCounty
}
\examples{
data(ExampleData)
}
\keyword{datasets}
