% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MapData}
\alias{MapData}
\title{Biological Records Centre Watsonian Vice County Boundaries}
\format{
A data frame with 1102774 rows and 7 variables
\describe {
 \item{long}{Longitude}
 \item{lat}{Latitude}
 \item{order}{Order}
 \item{hole}{Hole}
 \item{piece}{Piece}
 \item{id}{ID}
}
}
\source{
Biological Records Centre (https://github.com/BiologicalRecordsCentre/vice-counties)
}
\usage{
MapData
}
\description{
Shapefile containing vice county boundaries for Britain.
}
\examples{
data(MapData)
}
\keyword{datasets}
