% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesCheck.R
\name{OneSpecies}
\alias{OneSpecies}
\title{One Species List}
\usage{
OneSpecies(SpecimenColumn)
}
\arguments{
\item{SpecimenColumn}{The column where the specimen names are located in the dataset}
}
\value{
Dataframe OneSpeciesList with list of genera with only one species
}
\description{
Get a list of genera which only have one species for that genus
}
\examples{
OneSpecies(ExampleData$ScientificName)
}
