% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesCheck.R
\name{OneSpecimen}
\alias{OneSpecimen}
\title{One Specimen List}
\usage{
OneSpecimen(SpecimenColumn)
}
\arguments{
\item{SpecimenColumn}{The column where the specimen names are located in the dataset}
}
\value{
Dataframe OneSpecimenList with list of Species with only one specimen
}
\description{
Get a list of species which only have one specimen for that species.
}
\examples{
OneSpecimen(ExampleData$ScientificName)
}
