% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesCheck.R
\name{OrphanSpecies}
\alias{OrphanSpecies}
\title{Oprhan Species}
\usage{
OrphanSpecies(SpecimenColumn, Freq = 1)
}
\arguments{
\item{SpecimenColumn}{The column where the specimen names are located in the dataset}

\item{Freq}{The specified number of species in a genera. Default is 1.}
}
\value{
Dataframe OrphanSpeciesList with list of genera with with specified number of species.
}
\description{
Get a list of genera which have the specified number of species in.
}
\examples{
OrphanSpecies(ExampleData$ScientificName, 2)
}
