% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesCheck.R
\name{OrphanSpecimens}
\alias{OrphanSpecimens}
\title{Oprhan Specimens}
\usage{
OrphanSpecimens(SpecimenColumn, Freq = 1)
}
\arguments{
\item{SpecimenColumn}{The column where the specimen names are located in the dataset}

\item{Freq}{The specified number of specimens in a species Default is 1.}
}
\value{
Dataframe OrphanSpecimensList with list of species with with specified number of specimens
}
\description{
Get a list of species which have the specified number of specimens in.
}
\examples{
OrphanSpecimens(ExampleData$ScientificName, 2)
}
