% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectorsCheck.R
\name{SpeciesCollectors}
\alias{SpeciesCollectors}
\title{Species With One Collector}
\usage{
SpeciesCollectors(SpecimenColumn, ...)
}
\arguments{
\item{SpecimenColumn}{The column where the specimen names are located in the dataset}

\item{...}{All the columns which contain collector names, separated by (,)}
}
\value{
Dataframe OneCollector with list of species and collector name which only have one collector
}
\description{
Get a list of species which only have one collector.
}
\examples{
SpeciesCollectors(ExampleData$ScientificName, ExampleData$Collector1, ExampleData$Collector2, ExampleData$Collector3)
}
