% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViceCountyCheck.R
\name{SpecificDistributionMap}
\alias{SpecificDistributionMap}
\title{Species Distribution Map By Record Number With Census Data}
\usage{
SpecificDistributionMap(RecordNumber)
}
\arguments{
\item{Record}{The record number for the species in the data to produce map for.}
}
\value{
Plot of species distribution maps from collection data for specific record and plot of species distribution from census data.
}
\description{
Produce species distribution heat map for vice counties in Britain for a specific record in SuspectSpecies alongside distribution map from census data.
}
\examples{
SpecificDistributionMap(3)
}
