% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViceCountyCheck.R
\name{SpecificDistributionMapNoCensus}
\alias{SpecificDistributionMapNoCensus}
\title{Species Distribution Maps By Record Number Without Census Data}
\usage{
SpecificDistributionMapNoCensus(RecordNumber)
}
\arguments{
\item{Record}{The record number for the species in the data to produce map for.}
}
\value{
Plot of species distribution maps from collection data for specific record.
}
\description{
Produce species distribution heat map for vice counties in Britain for specific species from SuspectSpecies.
}
\examples{
SpecificDistributionMapNoCensus(3)
}
