% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViceCountyCheck.R
\name{SuspectSpeciesList}
\alias{SuspectSpeciesList}
\title{Suspect Species From Frequency}
\usage{
SuspectSpeciesList(SpecimenColumn, VCColumn)
}
\arguments{
\item{SpecimenColumn}{The column where the specimen names are located in the dataset}

\item{VCColumn}{The column where the vice county is located in the dataset}
}
\value{
Dataframe SuspectSpecies with list of suspect species
}
\description{
Get list of suspect species which have a frequency of 1 for any vice county they are found in. Species which have only one specimen for a vice county could be suspicious. Also allows you to view a dataframe: FilterZeros containing the frequency for every vice county for a species.
}
\examples{
SuspectSpeciesList(ExampleData$ScientificName, ExampleData$WatsonianViceCounty)
}
