% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViceCountyCheck.R
\name{UploadCensus}
\alias{UploadCensus}
\title{Upload Census Data}
\usage{
UploadCensus(CensusSpecies, CensusVC)
}
\arguments{
\item{CensusSpecies}{The column where the species names are located in the dataset}

\item{CensusVC}{The column where the vice counties are located in the dataset}
}
\value{
Usuable census data for distribution comparison
}
\description{
Use census data for any organism that uses Watsonian Vice Counties to produce distribution maps and compare collection and census data.
}
\examples{
UploadCensus(ExampleCensusData$Name, ExampleCensusData$VC_printed)
}
