% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectorsCheck.R
\name{SpeciesCollectors}
\alias{SpeciesCollectors}
\title{Species and Collectors}
\usage{
SpeciesCollectors(SpecimenColumn, NoCollectors = 1, ...)
}
\arguments{
\item{SpecimenColumn}{The column where the specimen names are located in the dataset
#' @param NoCollectors The specified number of collectors. Default is 1.}

\item{...}{All the columns which contain collector names, separated by (,)}
}
\value{
Dataframe GetCollectors with list of species and collector name for specified number of collectors.
}
\description{
Get a list of species which have specified number of collectors.
}
\examples{
SpeciesCollectors(ExampleData$ScientificName, ExampleData$Collector1, ExampleData$Collector2, ExampleData$Collector3, 2)
}
